requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
requests>=2.31.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
opencv-python-headless>=4.7.0
scikit-learn>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
